+++ packages/elysia/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vercel/elysia
 
+## 0.1.93
+
+### Patch Changes
+
+- @vercel/node@5.8.17
+
+## 0.1.92
+
+### Patch Changes
+
+- @vercel/node@5.8.16
+
 ## 0.1.91
 
 ### Patch Changes