+++ packages/express/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @vercel/express
 
+## 0.1.104
+
+### Patch Changes
+
+- @vercel/cervel@0.1.21
+
+## 0.1.103
+
+### Patch Changes
+
+- @vercel/cervel@0.1.20
+- @vercel/node@5.8.17
+
+## 0.1.102
+
+### Patch Changes
+
+- @vercel/node@5.8.16
+- @vercel/cervel@0.1.19
+
 ## 0.1.101
 
 ### Patch Changes