+++ packages/hono/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vercel/hono
 
+## 0.2.96
+
+### Patch Changes
+
+- @vercel/node@5.8.17
+
+## 0.2.95
+
+### Patch Changes
+
+- @vercel/node@5.8.16
+
 ## 0.2.94
 
 ### Patch Changes