+++ packages/koa/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vercel/koa
 
+## 0.1.76
+
+### Patch Changes
+
+- @vercel/node@5.8.17
+
+## 0.1.75
+
+### Patch Changes
+
+- @vercel/node@5.8.16
+
 ## 0.1.74
 
 ### Patch Changes